Front Home page, Shop page commit

+++ src/services/CategoryServices.ts
@@ -1,9 +1,9 @@
 import axios from "axios";
 import client from "../api/client";
 
-export const showCategoryService = () =>  client.get(`/categories`);
+export const showCategoryService = () =>  client.get(`/admin/categories`);
 
-export const showEditCategoryService = (categoryId: number) =>  client.get(`/categories/${categoryId}`);
+export const showEditCategoryService = (categoryId: number) =>  client.get(`/admin/categories/${categoryId}`);
 
 export const createCategoryService = (formData: any) => {
   const config = {
@@ -11,7 +11,7 @@ export const createCategoryService = (formData: any) => {
       "Content-Type": "multipart/form-data",
     },
   };
-  return client.post(`/categories`, formData, config);
+  return client.post(`/admin/categories`, formData, config);
 };
 
     export const editCategoryService = (categoryId: number, formData: any) => {
@@ -20,7 +20,7 @@ export const createCategoryService = (formData: any) => {
         "Content-Type": "multipart/form-data",
         },
     };
-    return client.put(`/categories/${categoryId}`, formData, config);
+    return client.put(`/admin/categories/${categoryId}`, formData, config);
     };
 
-export const deleteCategoryService = (categoryId: any) =>  client.delete(`/categories/${categoryId}`);
+export const deleteCategoryService = (categoryId: any) =>  client.delete(`/admin/categories/${categoryId}`);
