+++ src/services/ProductServices.ts
@@ -1,10 +1,10 @@
 import axios from "axios";
 import client from "../api/client";
 
-export const showProductService = () => client.get(`/products`);
+export const showProductService = () => client.get(`/admin/products`);
 
 export const showEditProductService = (productId: number) =>
-  client.get(`/products/${productId}`);
+  client.get(`/admin/products/${productId}`);
 
 export const createProductService = (formData: any) => {
   const config = {
@@ -12,7 +12,7 @@ export const createProductService = (formData: any) => {
       "Content-Type": "multipart/form-data",
     },
   };
-  return client.post(`/products`, formData, config);
+  return client.post(`/admin/products`, formData, config);
 };
 
 export const editProductService = (productId: number, formData: any) => {
@@ -21,11 +21,11 @@ export const editProductService = (productId: number, formData: any) => {
       "Content-Type": "multipart/form-data",
     },
   };
-  return client.put(`/products/${productId}`, formData, config);
+  return client.put(`/admin/products/${productId}`, formData, config);
 };
 
 export const deleteProductService = (productId: any) =>
-  client.delete(`/products/${productId}`);
+  client.delete(`/admin/products/${productId}`);
 
 export const uploadImageService = (formData: any) => {
   const config = {
@@ -33,12 +33,12 @@ export const uploadImageService = (formData: any) => {
       "Content-Type": "multipart/form-data",
     },
   };
-  return client.post(`/temp-images`, formData, config);
+  return client.post(`/admin/temp-images`, formData, config);
 };
 
 export const saveProductImagesService = (formData: FormData) => {
   console.log("save image called");
-  return client.post(`/save-product-images`, formData, {
+  return client.post(`/admin/save-product-images`, formData, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -46,14 +46,14 @@ export const saveProductImagesService = (formData: FormData) => {
 };
 
 export const deleteProductImageService = (imageId: number) => {
-  return client.delete(`/delete-product-image/${imageId}`);
+  return client.delete(`/admin/delete-product-image/${imageId}`);
 };
 
 export const makeDefaultProductImageService = (
   productId: number,
   image: string,
 ) => {
-  return client.get(`/change-product-default-image`, {
+  return client.get(`/admin/change-product-default-image`, {
     params: { image, product_id: productId },
   });
 };
